poly-autobetting/requirements.txt
httpx
python-dotenv
py-clob-client
py-builder-relayer-client
web3
eth-abi
websockets
python-telegram-bot>=20.0
